+++ src/main/kotlin/org/richinet/musicbackend/data/entity/TrackGroupId.kt
@@ -1,9 +1,8 @@
 package org.richinet.musicbackend.data.entity
 
 import java.io.Serializable
-import java.math.BigDecimal
 
 class TrackGroupId : Serializable {
-    var trackId: BigDecimal? = null
-    var groupId: BigDecimal? = null
+    var trackId: Long? = null
+    var groupId: Long? = null
 }
